feat: replace QueryMethodSelectionForm with EmailQueryForm for improved query handling and add custom scrollbar styles
